extract and rename `tool`, otherwise "tool" "util" ... all look too similar
