+++ frb_codegen/src/utils/dart_repository/dart_toolchain.rs
@@ -1,0 +1,49 @@
+use crate::commands::call_shell;
+use std::path::PathBuf;
+
+/// represents dart or flutter toolchain
+#[derive(Debug, PartialEq, Eq)]
+pub(crate) enum DartToolchain {
+    Dart,
+    Flutter,
+}
+
+impl ToString for DartToolchain {
+    fn to_string(&self) -> String {
+        match self {
+            DartToolchain::Dart => "dart",
+            DartToolchain::Flutter => "flutter",
+        }.to_string()
+    }
+}
+
+impl DartToolchain {
+    #[inline]
+    pub fn manifest_filename() -> &'static str {
+        "pubspec.yaml"
+    }
+
+    #[inline]
+    pub fn lock_filename() -> &'static str {
+        "pubspec.lock"
+    }
+
+    pub(crate) fn as_run_command(&self) -> Vec<PathBuf> {
+        match self {
+            DartToolchain::Dart => args!("dart"),
+            DartToolchain::Flutter => args!("flutter", "pub"),
+        }
+    }
+
+    pub(crate) fn available(&self) -> bool {
+        let toolchain = match self {
+            DartToolchain::Dart => "dart",
+            DartToolchain::Flutter => "flutter",
+        };
+        run!(call_shell[None], toolchain, "--version")
+            .unwrap()
+            .status
+            .success()
+    }
+}
+
